requirements.txt
typer
pytest
pypdf
python-docx
google-generativeai
